Improve building for windows

+++ lib/src/private/Pipelines/StaticObjectPipeline.cpp
@@ -132,8 +132,8 @@ VkResult CreateGraphicPipeline(VkDevice device,
     VkViewport viewport = {};
     viewport.x = 0.0f;
     viewport.y = 0.0f;
-    viewport.width = extent.width;
-    viewport.height = extent.height;
+    viewport.width = static_cast<float>(extent.width);
+    viewport.height = static_cast<float>(extent.height);
     viewport.minDepth = 0.0f;
     viewport.maxDepth = 1.0f;
